Create NumberGuessing Game
Project Title: Multiplayer Number Guessing Game in C

Description:
I, Mohd Uzaif Ansari, have developed a multiplayer number-guessing game in C where players must guess the correct number between 1 and 100. The goal of the game is for each player to guess the randomly generated number in the least number of attempts. The player with the fewest guesses wins, while the player with the most guesses loses. This fun and competitive game ensures that both strategy and luck are in play!

Project Date: 23rd September 2024